Check in 1.4.17

+++ src/lsiapi/lsiapihooks.cpp
@@ -332,7 +332,7 @@ int LsiApiHooks::runCallback(int level, int8_t *pEnableArray,
 //         pHook = find(pModule);
     }
 
-    lsi_hookinfo_t info = { this, pEnableArray, level, NULL };
+    lsi_hookinfo_t info = { this, pEnableArray, (int8_t)level, NULL };
     lsi_param_t param =
     {
         session,
